bugs fixed in modals, add task section, showing birthdays section

+++ database/migrations/2025_02_27_053014_create_daily_tasks_table.php
@@ -18,11 +18,11 @@ return new class extends Migration
             $table->unsignedBigInteger('project_id')->nullable();
             $table->string('project_name'); // New column
             $table->unsignedBigInteger('leave_id')->nullable(); // New column
-            $table->decimal('hours', 8, 2);
+            $table->unsignedInteger('hours'); // Changed from decimal to integer
             $table->text('task_description');
             $table->enum('task_status', ['pending', 'in_progress', 'completed']);
             $table->timestamps();
-
+        
             $table->foreign('project_id')->references('id')->on('projects')->onDelete('cascade');
             $table->foreign('leave_id')->references('id')->on('leaves')->onDelete('cascade');
         });
